+++ ios/RCTAgora/React/RCTAgoraRtcChannelModule.swift
@@ -14,11 +14,18 @@ class RCTAgoraRtcChannelModule: RCTEventEmitter {
     static let REACT_CLASS = "RCTAgoraRtcChannelModule"
 
     private var hasListeners = false
-    private lazy var manager: RtcChannelManager = {
-        return RtcChannelManager() { [weak self] methodName, data in
+    private var manager: RtcChannelManager?
+
+    override init() {
+        super.init()
+        manager = RtcChannelManager() { [weak self] methodName, data in
             self?.emit(methodName, data)
         }
-    }()
+    }
+
+    deinit {
+        manager?.Release()
+    }
 
     override class func moduleName() -> String! {
         return RCTAgoraRtcChannelModule.REACT_CLASS
@@ -26,10 +33,6 @@ class RCTAgoraRtcChannelModule: RCTEventEmitter {
 
     override func constantsToExport() -> [AnyHashable: Any]! {
         return ["prefix": RtcChannelEventHandler.PREFIX]
-    }
-
-    deinit {
-        manager.Release()
     }
 
     override class func requiresMainQueueSetup() -> Bool {
@@ -66,18 +69,18 @@ class RCTAgoraRtcChannelModule: RCTEventEmitter {
         return (bridge.module(for: RCTAgoraRtcEngineModule.classForCoder()) as? RCTAgoraRtcEngineModule)?.engine
     }
 
-    public func channel(_ channelId: String) -> AgoraRtcChannel? {
-        return manager[channelId]
+    func channel(_ channelId: String) -> AgoraRtcChannel? {
+        return manager?[channelId]
     }
 
     @objc func callMethod(_ methodName: String, _ params: NSDictionary?, _ resolve: RCTPromiseResolveBlock?, _ reject: RCTPromiseRejectBlock?) {
         if let `params` = params {
             if methodName == "create" {
-                params.setValue(engine(), forKey: "engine")
+                params.setValue(engine, forKey: "engine")
             }
-            manager.perform(NSSelectorFromString(methodName + "::"), with: params, with: PromiseCallback(resolve, reject))
+            manager?.perform(NSSelectorFromString(methodName + "::"), with: params, with: PromiseCallback(resolve, reject))
         } else {
-            manager.perform(NSSelectorFromString(methodName + ":"), with: PromiseCallback(resolve, reject))
+            manager?.perform(NSSelectorFromString(methodName + ":"), with: PromiseCallback(resolve, reject))
         }
     }
 }